Local Storage on Phone working
When you click on “Add” to add an assignment the log shows that a table
was created and when it tries to insert, it does so successfully.  Date
and Time are inserted as strings as well, however Time needs to be
parsed differently using a 24hr format. I also found a bug that crashes
the app when you rotate your phone. Other than that we need to work on
how to select stuff from the Table/storage and display it on the phone.